test(export): assert the exported quantity, and split the two failure stories
The probe test asserted not.toThrow(), which is satisfied by an exporter that
answers the version skew by declining to write quantities at all -- #2487's own
damage, a silent drop reported as success, arriving through the door opened to
fix it. It now reads the emitted IFCQUANTITYAREA.

declared-property-type.test.ts held two units with two different failure
stories: a real export driven over a parsed file, and a pure predicate over the
schema registry. AGENTS.md exempts test files from the ~400-line rule, so size
is not the reason -- the seam is. The halves share no imports at all: the unit
half needs no parser, no MutablePropertyView, no StepExporter and no IFC
fixture, and a failure in it is a statement about the predicate rather than
about the exporter. Test count is unchanged at 29.
